refactor prepareNestedLinkQb
