+++ packages/nocodb/src/db/IBaseModelSqlV2.ts
@@ -1,5 +1,10 @@
 import type { XKnex } from '~/db/CustomKnex';
-import type { NcApiVersion, NcContext } from 'nocodb-sdk';
+import type {
+  NcApiVersion,
+  NcContext,
+  NcRequest,
+  RelationTypes,
+} from 'nocodb-sdk';
 import type { Column, Model } from '~/models';
 import type { Knex } from 'knex';
 import type CustomKnex from '~/db/CustomKnex';
@@ -60,6 +65,20 @@ export interface IBaseModelSqlV2 {
       | string,
     alias?: string,
   ): string | Knex.Raw<any>;
+
+  afterAddChild(props: {
+    columnTitle: string;
+    columnId: string;
+    refColumnTitle: string;
+    rowId: unknown;
+    refRowId: unknown;
+    req: NcRequest;
+    model?: Model;
+    refModel?: Model;
+    displayValue: unknown;
+    refDisplayValue: unknown;
+    type: RelationTypes;
+  }): Promise<void>;
 
   get dbDriver(): CustomKnex;
   get isSqlite(): boolean;
